created file for resampling tests
